Fix issue with setting full CSP_DEFAULT_SRC variable
Our config library for getting values from the environment doesn't
allow for values that start or end with quotes (they're stripped)
so the value "'self'" needed in CSP doesn't work. You can have that
in the middle of a value, but not on the ends.

My workaround is to have the default list be immutable, and allow adding
extra values to the list via the CSP_DEFAULT_SRC env var.